+++ hardhat-config/index.js
@@ -1,0 +1,2 @@
+const {loadConfigFolder} = require('@animoca/ethereum-contracts/src/config');
+module.exports = loadConfigFolder(__dirname);