+++ hardhat.config.js
@@ -1,3 +1,5 @@
-// WARNING: this is a template file, do not edit. Edit hardhat.config.local.js instead.
-require('@animoca/ethereum-hardhat-bootstrap/hardhat-plugins');
-module.exports = require('@animoca/ethereum-hardhat-bootstrap').loadConfig();
+const {mergeConfigs} = require('@animoca/ethereum-contracts/src/config');
+
+require('@animoca/ethereum-contracts/hardhat-plugins');
+
+module.exports = mergeConfigs(require('@animoca/ethereum-contracts/hardhat-config'), require('./hardhat-config'));