+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 6.0.0
+
+### Breaking changes
+
+- Migrated to the new core library `@animoca/ethereum-contracts`, replacing `@animoca/ethereum-contracts-core`.
+- Moved the old artifacts folders under `artifacts_previous`.
+- Revisited the fusion mechanism.
+
+### New features
+
+- Added `REVVMotorsportVouchers` contract.
+- Added `REVVMotorsportShard` contract.
+- Added `REVVRacingCatalyst` contract.
+- Added `REVVRacingNFT` contract.
+
 ## 5.1.0
 
 ### Chores